message_service_api/requirements.txt
altgraph==0.10.2
appdirs==1.4.0
APScheduler==2.0.3
backports-abc==0.5
bdist-mpkg==0.5.0
certifi==2017.1.23
click==6.7
Flask==0.12
flask-mongoengine==0.9.2
Flask-Redis==0.3.0
Flask-SocketIO==2.8.4
Flask-SQLAlchemy==2.1
Flask-WTF==0.14.2
itsdangerous==0.24
Jinja2==2.9.5
kafka-python==1.3.2
kazoo==2.2.1
macholib==1.5.1
MarkupSafe==0.23
modulegraph==0.10.4
mongoengine==0.11.0
MySQL-python==1.2.5
packaging==16.8
py==1.4.32
pycrypto==2.6.1
pymongo==3.4.0
pyOpenSSL==0.13.1
pyparsing==2.1.10
pytest==3.0.6
python-dateutil==1.5
python-engineio==1.2.3
python-socketio==1.7.1
redis==2.10.5
requests==2.13.0
singledispatch==3.4.0.3
six==1.10.0
SQLAlchemy==1.1.5
tabulate==0.7.7
tornado==4.4.2
uWSGI==2.0.14
Werkzeug==0.11.15
WTForms==2.1
WTForms-JSON==0.3.1
xattr==0.6.4
zope.interface==4.1.1
flask-limiter==0.9.3
limits==1.2.1
